Refactor : Handler, HandlerTest 클래스 위치 변경
