+++ sport_companion/module-api/src/main/java/com/service/sport_companion/api/component/club/SeasonHandler.java
@@ -1,4 +1,4 @@
-package com.service.sport_companion.api.component;
+package com.service.sport_companion.api.component.club;
 
 import com.service.sport_companion.domain.entity.SeasonsEntity;
 import com.service.sport_companion.domain.repository.SeasonsRepository;
