+++ sport_companion/module-api/src/test/java/com/service/sport_companion/api/component/club/FixtureHandlerTest.java
@@ -1,4 +1,4 @@
-package com.service.sport_companion.api.component;
+package com.service.sport_companion.api.component.club;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.when;
